requirements.txt
pybars3
pystache
pyyaml
n2w
stringcase